+++ src/collectors/network/network.py
@@ -13,7 +13,6 @@ using /proc/net/dev.
 import diamond.collector
 from diamond.collector import str_to_bool
 import diamond.convertor
-from diamond.pycompat import long
 import os
 import re
 
@@ -115,7 +114,7 @@ class NetworkCollector(diamond.collector.Collector):
                 metric_name = '.'.join([device, s])
                 # Get Metric Value
                 metric_value = self.derivative(metric_name,
-                                               long(v),
+                                               int(v),
                                                diamond.collector.MAX_COUNTER)
 
                 # Convert rx_bytes and tx_bytes
