+++ src/collectors/proc/proc.py
@@ -13,7 +13,6 @@ The ProcessStatCollector collects metrics on process stats from
 import platform
 import os
 import diamond.collector
-from diamond.pycompat import long
 # Detect the architecture of the system
 # and set the counters for MAX_VALUES
 # appropriately. Otherwise, rolling over
@@ -64,7 +63,7 @@ class ProcessStatCollector(diamond.collector.Collector):
                 metric_name = data[0]
                 metric_value = int(data[1])
                 metric_value = int(self.derivative(metric_name,
-                                                   long(metric_value),
+                                                   int(metric_value),
                                                    counter))
                 self.publish(metric_name, metric_value)
